splitt DataElement ut i egne filer
