+++ frontend/mr-admin-flate/src/components/tilsagn/beregning/TilsagnRegnestykke.tsx
@@ -1,6 +1,7 @@
 import { CalculationDto } from "@mr/api-client-v2";
 import { HStack } from "@navikt/ds-react";
-import { DataDrivenTable, renderCell } from "@/components/tabell/DataDrivenTable";
+import { DataDrivenTable } from "@/components/tabell/DataDrivenTable";
+import { getDataElement } from "@/components/data-element/DataElement";
 
 interface Props {
   regnestykke: CalculationDto;
@@ -10,7 +11,7 @@ export function TilsagnRegnestykke({ regnestykke }: Props) {
   const expression = regnestykke.expression ? (
     <HStack gap="2">
       {regnestykke.expression.map((entry, idx) => (
-        <span key={idx}>{renderCell(entry)} </span>
+        <span key={idx}>{getDataElement(entry)} </span>
       ))}
     </HStack>
   ) : null;
